InputStateHandler renamed to GameInputHandler.
